move influx to new source file
