+++ influx.hpp
@@ -1,0 +1,1 @@
+int post_data(const properties *);
